Writing checksums as ints to save 4 bytes/block.

+++ src/main/java/com/github/batkinson/jrsync/Metadata.java
@@ -45,7 +45,7 @@ public class Metadata {
         for (int i = 0; i<wholeBlocks; i++) {
             source.readFully(block);
             checksum.update(block);
-            metadata.writeLong(checksum.getValue());
+            metadata.writeInt((int)checksum.getValue());
             metadata.write(blockDigest.digest(block));
             fileDigest.update(block);
         }
